gradle: Add Bndrun task type
The TestOSGi task type is also refactored to extend the new Bndrun task
type.
