+++ biz.aQute.bnd.gradle/src/aQute/bnd/gradle/TestOSGi.groovy
@@ -19,10 +19,10 @@
  * <li>ignoreFailures - If true the task will not fail if the any
  * test cases fail. Otherwise, the task will fail is any test case
  * fails. The default is false.</li>
- * <li>bndrun - This is the bndrun file to be resolved.
+ * <li>bndrun - This is the bndrun file to be tested.
  * This property must be set.</li>
  * <li>workingDir - This is the directory for the test case execution.
- * The default for destinationDir is temporaryDir.</li>
+ * The default for workingDir is temporaryDir.</li>
  * <li>bundles - This is the collection of files to use for locating
  * bundles during the bndrun execution. The default is
  * 'sourceSets.main.runtimeClasspath' plus
@@ -39,36 +39,16 @@ package aQute.bnd.gradle
 
 import static aQute.bnd.gradle.BndUtils.logReport
 
-import aQute.bnd.build.Run
 import aQute.bnd.build.Project
-import aQute.bnd.build.Workspace
-import aQute.bnd.service.RepositoryPlugin
 
-import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.options.Option
 
-public class TestOSGi extends DefaultTask {
-  /**
-   * Whether test execution failures should be ignored.
-   *
-   * <p>
-   * If <code>true</code>, then test execution failures will not fail the task.
-   * Otherwise, a test execution failure will fail the task. The default is
-   * <code>false</code>.
-   */
-  @Input
-  boolean ignoreFailures
-
-  private File workingDir
-  private File bndrun
+public class TestOSGi extends Bndrun {
   private List<String> tests
-  private final Workspace bndWorkspace
 
   /**
    * Create a TestOSGi task.
@@ -76,51 +56,6 @@ public class TestOSGi extends DefaultTask {
    */
   public TestOSGi() {
     super()
-    bndWorkspace = project.findProperty('bndWorkspace')
-    ignoreFailures = false
-    workingDir = temporaryDir
-    if (bndWorkspace == null) {
-      convention.plugins.bundles = new FileSetRepositoryConvention(this)
-    }
-  }
-
-  /**
-   * Return the bndrun file for the test execution.
-   *
-   */
-  @InputFile
-  public File getBndrun() {
-    return bndrun
-  }
-
-  /**
-   * Set the bndfile for the test execution.
-   *
-   * <p>
-   * The argument will be handled using
-   * Project.file().
-   */
-  public void setBndrun(Object file) {
-    bndrun = project.file(file)
-  }
-
-  /**
-   * Return the working dir for the test execution.
-   *
-   */
-  public File getWorkingDir() {
-    return workingDir
-  }
-
-  /**
-   * Set the working dir for the test execution.
-   *
-   * <p>
-   * The argument will be handled using
-   * Project.file().
-   */
-  public void setWorkingDir(Object dir) {
-    workingDir = project.file(dir)
   }
 
   /**
@@ -150,44 +85,8 @@ public class TestOSGi extends DefaultTask {
       return tests;
   }
 
-  /**
-   * Test the bndrun file.
-   *
-   */
-  @TaskAction
-  void testOSGi() {
-    Workspace workspace = bndWorkspace
-    if (workspace != null && bndrun == project.bnd.project.getPropertiesFile()) {
-      testWorker(project.bnd.project)
-      return
-    }
-    Run.createRun(workspace, bndrun).withCloseable { Run run ->
-      Workspace runWorkspace = run.getWorkspace()
-      project.mkdir(workingDir)
-      run.setBase(workingDir)
-      if (run.isStandalone()) {
-        runWorkspace.setOffline(workspace != null ? workspace.isOffline() : project.gradle.startParameter.offline)
-        File cnf = new File(temporaryDir, Workspace.CNFDIR)
-        project.mkdir(cnf)
-        runWorkspace.setBuildDir(cnf)
-        if (convention.findPlugin(FileSetRepositoryConvention)) {
-          runWorkspace.addBasicPlugin(getFileSetRepository(name))
-          for (RepositoryPlugin repo : runWorkspace.getRepositories()) {
-            repo.list(null)
-          }
-        }
-      }
-      run.getInfo(runWorkspace)
-      logReport(run, logger)
-      if (!run.isOk()) {
-        throw new GradleException("${run.getPropertiesFile()} workspace errors")
-      }
-
-      testWorker(run)
-    }
-  }
-
-  void testWorker(Project run) {
+  @Override
+  protected void worker(Project run) {
     try {
       logger.info 'Running tests for {} in {}', run.getPropertiesFile(), run.getBase()
       run.test(resultsDir, tests);
